# backend/docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  github_com_0ScPro0_affiliate-system_internal_core_transport_dto.PartnerResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UpdatePartnerRequest:
    properties:
      description:
        maxLength: 1000
        minLength: 1
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - id
    - name
    type: object
  github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UpdateUserRequest:
    properties:
      email:
        maxLength: 100
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      password_hash:
        maxLength: 255
        minLength: 1
        type: string
      username:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - id
    type: object
  github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UserResponse:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
      username:
        type: string
    type: object
  github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  description: Affiliate System API
  title: Affiliate System API
  version: "1.0"
paths:
...
  /partners/{id}:
    delete:
      consumes:
      - application/json
      description: Delete partner by ID
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Partner deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete partner
      tags:
      - partners
    get:
      consumes:
      - application/json
      description: Get partner by its unique identifier
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Partner found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.PartnerResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      summary: Get partner by ID
      tags:
      - partners
    patch:
      consumes:
      - application/json
      description: Update existing partner by ID
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdatePartner request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UpdatePartnerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Partner updated successfully
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.PartnerResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update partner
      tags:
      - partners
...
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: User deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user by its unique identifier
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update existing user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateUser request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_dto.UserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_0ScPro0_affiliate-system_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by a space and your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"